requirements.txt
Django==1.8.3
djangorestframework==3.2.0
psycopg2==2.6.1
